// Frontend/src/models/Types/sizeCalculator/DTO/SizeDTO.ts
export type SizeDTO = {
    TShoulderSizeF: number | null;
    TChestSizeF: number | null;
    TWaistSizeF: number | null; 
    TShoulderSizeM: number | null;
    TChestSizeM: number | null;
    TWaistSizeM: number | null; 
    OSShoulderSize: number | null;
    OSChestSize: number | null;
    OSWaistSize: number | null; 
}